// src/app/models/vino.model.ts
export class Vino {
  Alcohol: number;
  Chlorides: number;
  CitricAcid: number;
  Density: number;
  FixedAcidity: number;
  FreeSulfurDioxide: number;
  Id: number;
  IdProductor: number;
  Nombre: string;
  PH: number;
  Quality: number;
  Redwine: boolean;
  ResidualSugar: number;
  Sulphates: number;
  TotalSulfurDioxide: number;
  VolatileAcidity: number;
}